backend/requirements.txt
anthropic==0.71.0
blinker==1.9.0
boto3==1.40.69
botocore==1.40.69
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyMySQL==1.1.2
pyScss==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3

# ML dependencies for fine-tuned QWEN model
transformers>=4.36.0
torch>=2.0.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.14.0
sentencepiece>=0.1.99
huggingface-hub>=0.19.0
